+++ src/macros.rs
@@ -85,3 +85,29 @@ macro_rules! heprintln {
         $crate::export::hstderr_fmt(format_args!(concat!($s, "\n"), $($tt)*))
     };
 }
+
+/// Macro that prints and returns the value of a given expression
+/// for quick and dirty debugging. Works exactly like `dbg!` in
+/// the standard library, replacing `eprintln` with `heprintln`.
+#[macro_export]
+macro_rules! dbg {
+    () => {
+        $crate::hprintln!("[{}:{}]", file!(), line!());
+    };
+    ($val:expr) => {
+        // Use of `match` here is intentional because it affects the lifetimes
+        // of temporaries - https://stackoverflow.com/a/48732525/1063961
+        match $val {
+            tmp => {
+                $crate::hprintln!("[{}:{}] {} = {:#?}",
+                    file!(), line!(), stringify!($val), &tmp);
+                tmp
+            }
+        }
+    };
+    // Trailing comma with single argument is ignored
+    ($val:expr,) => { $crate::dbg!($val) };
+    ($($val:expr),+ $(,)?) => {
+        ($($crate::dbg!($val)),+,)
+    };
+}